+++ src/code/platform/lookup/electron/lookupPaletteService.ts
@@ -1,8 +1,9 @@
 import { globalShortcut } from "electron";
 import { ILogService } from "src/base/common/logger";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
-import { IWindowCreationOptions, IWindowInstance } from "src/code/platform/window/common/window";
+import { IWindowCreationOptions } from "src/code/platform/window/common/window";
 import { IMainWindowService } from "src/code/platform/window/electron/mainWindowService";
+import { IWindowInstance } from "src/code/platform/window/electron/windowInstance";
 
 const GLOBAL_LOOKUP_KEY = 'Control+Shift+F';
