+++ src/code/platform/window/electron/mainWindowService.ts
@@ -8,8 +8,8 @@ import { createDecorator } from "src/code/platform/instantiation/common/decorato
 import { IInstantiationService } from "src/code/platform/instantiation/common/instantiation";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
 import { IMainLifeCycleService } from "src/code/platform/lifeCycle/electron/mainLifeCycleService";
-import { ToOpenType, IUriToOpenConfiguration, IWindowConfiguration, IWindowCreationOptions, IWindowInstance } from "src/code/platform/window/common/window";
-import { WindowInstance } from "src/code/platform/window/electron/windowInstance";
+import { ToOpenType, IUriToOpenConfiguration, IWindowConfiguration, IWindowCreationOptions } from "src/code/platform/window/common/window";
+import { IWindowInstance, WindowInstance } from "src/code/platform/window/electron/windowInstance";
 import { URI } from "src/base/common/file/uri";
 
 export const IMainWindowService = createDecorator<IMainWindowService>('main-window-service');
@@ -27,6 +27,11 @@ export interface IMainWindowService extends Disposable {
      * @description Returns all the running windows.
      */
     windows(): ReadonlyArray<IWindowInstance>;
+
+    /**
+     * @description Returns the opened window by the given id.
+     */
+    getWindowByID(id: number): IWindowInstance | undefined;
 
     /**
      * @description Returns the number of running window.
@@ -73,6 +78,15 @@ export class MainWindowService extends Disposable implements IMainWindowService 
 
     public windows(): readonly IWindowInstance[] {
         return this._windows;
+    }
+
+    public getWindowByID(id: number): IWindowInstance | undefined {
+        for (const window of this._windows) {
+            if (window.id === id) {
+                return window;
+            }
+        }
+        return undefined;
     }
 
     // [public methods]
